Modify PATH to appease Homebrew
On new install using the wellmatch_dev_setup script, homebrew installs correctly, but gives a warning after running `brew doctor`.  This is because its path is not loaded before /usr/bin, meaning its installed packages would not override system defaults.
This adds the line that Homebrew expects us to.  Not sure if this works for those using zsh.