Swap files only stored in common directory
* Rather than allowing swap files to get stored in the current directory
we want to force them to be written to a common directory stored outside
the current repo. This updates the symlinked .vim to hold a .swp where
these files will be stored. It also includes an update to our .gitignore
to exclude any other contents of the directory.
* We don't seem to be using vim backups and there's a lot of support out
there for disabling them. Rather than manage an additional common
directory for those, we're explicitly disabling the feature.