Export DOTFILES_DIR to enable editing of dotfiles with vim
Modify .vimrc to no longer attempt to use the repo version of the .vimrc
when trying to edit dotfiles.  To test changes, simply edit your .vimrc
as usual, source .vimrc, and restart vim.

This fixes behavior that caused vim to hang when attempting to edit
files within our repo.

In addition, the path to the DOTFILES is now available for any other
scripting needs we might have.