Run tests from vim within containers
This removes support for the typical local workflow.  While I'm not
entirely happy with this, it fixes a huge pain point for most of us.

An additional PR could ensure that we only do this when an ENV variable
is present, indicating that we're on a pair box.

Developers working locally (without containers) will be (sadpanda).