Add vim-vinegar Plugin
vim-vinegar is a navigation plugin that I think we could get some
benefit out of. I find it quicker to move around a directory tree using
it than with Nerdtree. Since we have a bunch of other tpope plugins, I figured
this would be a nice addition to the list.