Remove Git Commit Vim Colorcolumns
We have the vim-fugitive plugin in our workflow. This plugin
automatically forces a line break when giving a detailed description to
a git commit, and it changes the text color when your git commit header
is too long. The colorcolumns here feel like they may be noise when
trying to write a commit message.