Prefer jq over python for JSON formatting
This introduces a new dependency, but is able to format JSON correctly, i.e. without some of the side-effects of the python implementation.

First, it respects alphabetical order. A formatting tool should change the shape of the data, not its content.  Or, to be more specific, a diff
should display only whitespace changes.

Second, it formats with 2 spaces by default. I tried fixing this with the python tool, but was unable to.

Third, this new dependency is badass in its own right, and deserves to be on our systems regardless.  It is super fast (written in C), and meant to be
a good UNIX citizen.