Fix vimbundle update process
Previously, additions to .vimbundle.local would fail because the symlink used when running /bin/upgrade
would not be updated. Given that .vimbundle.local was only meant to be temporary, we have removed it in
favor of the normal .vimbundle file.

Also, fix $WELLMATCH_DIR to point at ~/src/ for EC2 pairboxes.